Move {Clear,Exit}Command back into exec
This is to clean up the repl module so that it doesn't include commands,
which in turn will make it easier to move it to the cli.

Note that only ExitCommand should move to make this possible, but this and
ClearCommand are very similar in that they only affect the state of the
machine and thus belong in the same place.
